move Athena components to Athena namespace
