+++ catalog/app/containers/Bucket/Queries/Athena/QueryEditor.tsx
