+++ catalog/app/containers/Bucket/Queries/Athena/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Athena'
